factored out imperative assignment operators from generic mathematical operators
