+++ lang/c/c.talon
@@ -4,10 +4,13 @@ mode: command
 and mode: user.auto_lang
 and code.language: c
 -
+tag(): user.code_imperative
 tag(): user.code_operators
+tag(): user.code_operators_imperative
 tag(): user.code_comment
 tag(): user.code_block_comment
-tag(): user.code_generic
+tag(): user.code_gui_functions
+tag(): user.code_gui_libraries
 settings():
     user.code_private_function_formatter = "SNAKE_CASE"
     user.code_protected_function_formatter = "SNAKE_CASE"
@@ -17,8 +20,8 @@ settings():
     user.code_public_variable_formatter = "SNAKE_CASE"
     # whether or not to use uint_8 style datatypes
     #    user.use_stdint_datatypes = 1
-    
-    
+
+
 
 ^funky <user.text>$: user.code_default_function(text)
 ^static funky <user.text>$: user.code_private_static_function(text)
@@ -50,16 +53,16 @@ push brackets:
     key(enter)
     key(enter)
     edit.up()
-    
+
 # Declare variables or structs etc.
 # Ex. * int myList
 <user.c_variable> <phrase>:
     insert("{c_variable} ")
     insert(user.formatted_text(phrase, "PRIVATE_CAMEL_CASE,NO_SPACES"))
-    
+
 <user.c_variable> <user.letter>:
     insert("{c_variable} {letter} ")
-    
+
 # Ex. (int *)
 cast to <user.c_cast>: "{c_cast}"
 standard cast to <user.stdint_cast>: "{stdint_cast}"
@@ -70,7 +73,7 @@ standard <user.stdint_types>: "{stdint_types}"
 int main:
     insert("int main()")
     edit.left()
-    
+
 toggle includes: user.code_toggle_libraries()
 include <user.code_libraries>:
     user.code_insert_library(code_libraries, "")
