+++ lang/csharp/csharp.talon
@@ -4,9 +4,11 @@ mode: command
 and mode: user.auto_lang
 and code.language: csharp
 -
+tag(): user.code_imperative
 tag(): user.code_operators
+tag(): user.code_operators_imperative
 tag(): user.code_comment
-tag(): user.code_generic
+tag(): user.code_gui_functions
 settings():
     user.code_private_function_formatter = "PRIVATE_CAMEL_CASE"
     user.code_protected_function_formatter = "PUBLIC_CAMEL_CASE"
@@ -14,7 +16,7 @@ settings():
     user.code_private_variable_formatter = "PRIVATE_CAMEL_CASE"
     user.code_protected_variable_formatter = "PUBLIC_CAMEL_CASE"
     user.code_public_variable_formatter = "PUBLIC_CAMEL_CASE"
-    
+
 
 ^funky <user.text>$: user.code_default_function(text)
 ^pro funky <user.text>$: user.code_protected_function(text)
