+++ lang/java/java.talon
@@ -2,10 +2,12 @@ mode: user.java
 mode: user.auto_lang
 and code.language: java
 -
+tag(): user.code_imperative
 tag(): user.code_operators
+tag(): user.code_operators_imperative
 tag(): user.code_comment
 tag(): user.code_block_comment
-tag(): user.code_generic
+tag(): user.code_gui_functions
 
 settings():
     user.code_private_function_formatter = "PRIVATE_CAMEL_CASE"
@@ -16,7 +18,7 @@ settings():
     user.code_public_variable_formatter = "PRIVATE_CAMEL_CASE"
 
 # Types Commands
-boxed [type] {user.java_boxed_types}: 
+boxed [type] {user.java_boxed_types}:
     insert(user.java_boxed_types)
     key("space")
 
@@ -30,13 +32,13 @@ type {user.code_type} array:
     insert(user.code_type)
     user.code_operator_subscript()
 
-[state] {user.java_access_modifiers}: 
+[state] {user.java_access_modifiers}:
     insert(user.java_access_modifiers)
     key("space")
 
-[state] {user.java_modifiers}: 
+[state] {user.java_modifiers}:
     insert(user.java_modifiers)
-    key("space") 
+    key("space")
 
 op array:
     user.code_operator_subscript()
@@ -44,7 +46,7 @@ op array:
 op new:
     insert("new ")
 
-# Methods    
+# Methods
 ^method <user.text>$: user.code_default_function(text)
 ^pro method <user.text>$: user.code_protected_function(text)
 ^pub method <user.text>$: user.code_public_function(text)
