+++ lang/r/r.talon
@@ -6,9 +6,12 @@ and code.language: r
 -
 # TODO: functions
 
+tag(): user.code_imperative
 tag(): user.code_operators
+tag(): user.code_operators_imperative
 tag(): user.code_comment
-tag(): user.code_generic
+tag(): user.code_gui_functions
+tag(): user.code_gui_libraries
 settings():
     user.code_private_function_formatter = "SNAKE_CASE"
     user.code_protected_function_formatter = "SNAKE_CASE"
@@ -16,6 +19,7 @@ settings():
     user.code_private_variable_formatter = "SNAKE_CASE"
     user.code_protected_variable_formatter = "SNAKE_CASE"
     user.code_public_variable_formatter = "SNAKE_CASE"
+
 toggle library: user.code_toggle_libraries()
 library <user.code_libraries>:
     user.code_insert_library(code_libraries, "")
@@ -28,7 +32,7 @@ library <user.code_libraries>:
     key(enter)
 state na:
     insert("NA")
-    
+
 ^function define <user.text>$: user.code_private_function(text)
 
 named arg {user.code_parameter_name}: user.code_insert_named_argument(code_parameter_name)
