+++ lang/tags/gui/libraries.talon
@@ -1,0 +1,8 @@
+tag: user.code_gui_libraries
+-
+# NOTE: This file does not define any commands, as the commands vary from
+#       language to language, e.g., Python uses 'import', C uses 'include',
+#       R uses 'library', etcetera.
+
+# TODO: If this ever becomes possible, we should abstract over these commands
+#       using a variable which can be set to the context-specific word.
