+++ lang/tags/imperative.talon
@@ -39,18 +39,6 @@ state next: user.code_next()
 state true: user.code_true()
 state false: user.code_false()
 
-# show and print functions and libraries
-toggle funk: user.code_toggle_functions()
-funk <user.code_functions>:
-    user.code_insert_function(code_functions, "")
-funk cell <number>:
-    user.code_select_function(number - 1, "")
-funk wrap <user.code_functions>:
-    user.code_insert_function(code_functions, edit.selected_text())
-funk wrap <number>:
-    user.code_select_function(number - 1, edit.selected_text())
-dock string: user.code_document_string()
-
 # for annotating function parameters
 is type {user.code_type}: user.code_insert_type_annotation(code_type)
 returns [type] {user.code_type}: user.code_insert_return_type(code_type)
