+++ lang/tags/operators.py
@@ -3,12 +3,14 @@ from talon import Context, Module
 ctx = Context()
 mod = Module()
 
-# TODO: rename this tag to 'code_operator'
+# TODO: rename this tag to 'code_operator' for consistency?
 mod.tag("code_operators", desc="Tag for enabling generic operator commands")
 
 
 @mod.action_class
 class Actions:
+
+    # TODO: move into pointer tag?
 
     def code_operator_indirection():
         """code_operator_indirection"""
@@ -25,26 +27,14 @@ class Actions:
     def code_operator_subscript():
         """code_operator_subscript (e.g., C++ [])"""
 
-    def code_operator_assignment():
-        """code_operator_assignment"""
-
     def code_operator_subtraction():
         """code_operator_subtraction"""
-
-    def code_operator_subtraction_assignment():
-        """code_operator_subtraction_equals"""
 
     def code_operator_addition():
         """code_operator_addition"""
 
-    def code_operator_addition_assignment():
-        """code_operator_addition_assignment"""
-
     def code_operator_multiplication():
         """code_operator_multiplication"""
-
-    def code_operator_multiplication_assignment():
-        """code_operator_multiplication_assignment"""
 
     def code_operator_exponent():
         """code_operator_exponent"""
@@ -52,14 +42,8 @@ class Actions:
     def code_operator_division():
         """code_operator_division"""
 
-    def code_operator_division_assignment():
-        """code_operator_division_assignment"""
-
     def code_operator_modulo():
         """code_operator_modulo"""
-
-    def code_operator_modulo_assignment():
-        """code_operator_modulo_assignment"""
 
     def code_operator_equal():
         """code_operator_equal"""
@@ -91,32 +75,14 @@ class Actions:
     def code_operator_bitwise_and():
         """code_operator_bitwise_and"""
 
-    def code_operator_bitwise_and_assignment():
-        """code_operator_and"""
-
-    def code_operator_increment():
-        """code_operator_increment"""
-
     def code_operator_bitwise_or():
         """code_operator_bitwise_or"""
-
-    def code_operator_bitwise_or_assignment():
-        """code_operator_or_assignment"""
 
     def code_operator_bitwise_exclusive_or():
         """code_operator_bitwise_exclusive_or"""
 
-    def code_operator_bitwise_exclusive_or_assignment():
-        """code_operator_bitwise_exclusive_or_assignment"""
-
     def code_operator_bitwise_left_shift():
         """code_operator_bitwise_left_shift"""
 
-    def code_operator_bitwise_left_shift_assignment():
-        """code_operator_bitwise_left_shift_assigment"""
-
     def code_operator_bitwise_right_shift():
         """code_operator_bitwise_right_shift"""
-
-    def code_operator_bitwise_right_shift_assignment():
-        """code_operator_bitwise_right_shift_assignment"""
