+++ lang/talon/talon.talon
@@ -6,7 +6,8 @@ and code.language: talon
 -
 tag(): user.code_operators
 tag(): user.code_comment
-# uncomment user.talon_populate_lists tag to activate talon-specific lists of actions, scopes, modes etcetera. 
+tag(): user.code_gui_functions
+# uncomment user.talon_populate_lists tag to activate talon-specific lists of actions, scopes, modes etcetera.
 # Do not enable this tag with dragon, as it will be unusable.
 # with conformer, the latency increase may also be unacceptable depending on your cpu
 # see https://github.com/knausj85/knausj_talon/issues/600
@@ -20,7 +21,7 @@ action block:
     edit.left()
 setting block:
     insert("settings():\n\t")
-setting {user.talon_settings}: 
+setting {user.talon_settings}:
     user.paste("{talon_settings} = ")
 #context requirements
 win require:
@@ -31,16 +32,16 @@ linux require:
     insert("os: linux\n")
 title require:
     insert("win.title: ")
-application [require] [{user.talon_apps}]: 
+application [require] [{user.talon_apps}]:
     app = talon_apps or ""
     user.paste("app: {app}")
-mode require [{user.talon_modes}]: 
+mode require [{user.talon_modes}]:
     mode = talon_modes or ""
     user.paste("mode: {mode}")
-tag require [{user.talon_tags}]: 
+tag require [{user.talon_tags}]:
     tag = talon_tags or ""
     user.paste("tag: {tag}")
-tag set [{user.talon_tags}]: 
+tag set [{user.talon_tags}]:
     tag = talon_tags or ""
     user.paste("tag(): {tag}")
 # requires user.talon_populate_lists tag. do not use with dragon
