+++ lang/vimscript/vimscript.talon
@@ -4,9 +4,10 @@ mode: command
 and mode: user.auto_lang
 and code.language: vimscript
 -
+tag(): user.code_imperative
 tag(): user.code_operators
+tag(): user.code_operators_imperative
 tag(): user.code_comment
-tag(): user.code_generic
 # XXX - revisit these
 settings():
     user.code_private_function_formatter = "SNAKE_CASE"
@@ -23,15 +24,15 @@ assign [<user.vimscript_scope>] (variable|var) [<user.text>] [over]:
     insert("let ")
     insert(vimscript_scope or '')
     user.code_private_variable_formatter(text)
-    
+
 [<user.vimscript_scope>] (variable|var) [<user.text>] [over]:
     insert(vimscript_scope or '')
     user.code_private_variable_formatter(text)
-    
+
 # see lang/vimscript/vimscript.py for list
 <user.vimscript_functions>:
     insert("{vimscript_functions} ")
-    
+
 # XXX - possibly overlap with some programming.talon
 state command: "command! "
 state end if: "endif"
